move type Etag to its own file
